# backend/tests/openapi_snapshot.json
{
  "components": {
    "schemas": {
...
      "CatalogTeamOut": {
        "properties": {
          "abbreviation": {
            "title": "Abbreviation",
            "type": "string"
          },
          "color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          },
          "logo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logo Url"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "provider_key": {
            "title": "Provider Key",
            "type": "string"
          }
        },
        "required": [
          "provider_key",
          "name",
          "abbreviation"
        ],
        "title": "CatalogTeamOut",
        "type": "object"
      },
      "CatalogTeamsOut": {
        "properties": {
          "league_id": {
            "title": "League Id",
            "type": "string"
          },
          "teams": {
            "items": {
              "$ref": "#/components/schemas/CatalogTeamOut"
            },
            "title": "Teams",
            "type": "array"
          }
        },
        "required": [
          "league_id",
          "teams"
        ],
        "title": "CatalogTeamsOut",
        "type": "object"
      },
...
      "FollowRequest": {
        "properties": {
          "selections": {
            "items": {
              "$ref": "#/components/schemas/FollowSelection"
            },
            "title": "Selections",
            "type": "array"
          }
        },
        "required": [
          "selections"
        ],
        "title": "FollowRequest",
        "type": "object"
      },
      "FollowSelection": {
        "properties": {
          "follow_all": {
            "default": false,
            "title": "Follow All",
            "type": "boolean"
          },
          "league_id": {
            "title": "League Id",
            "type": "string"
          },
          "team_provider_keys": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Team Provider Keys",
            "type": "array"
          }
        },
        "required": [
          "league_id"
        ],
        "title": "FollowSelection",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LeagueOut": {
        "properties": {
          "follow_all": {
            "default": false,
            "title": "Follow All",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "sport": {
            "title": "Sport",
            "type": "string"
          }
        },
        "required": [
          "id",
          "sport",
          "name",
          "provider"
        ],
        "title": "LeagueOut",
        "type": "object"
      },
...
      "TeamOut": {
        "properties": {
          "abbreviation": {
            "title": "Abbreviation",
            "type": "string"
          },
          "color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "league_id": {
            "title": "League Id",
            "type": "string"
          },
          "logo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logo Url"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "league_id",
          "name",
          "abbreviation"
        ],
        "title": "TeamOut",
        "type": "object"
      },
...
      "TeamsOut": {
        "properties": {
          "leagues": {
            "items": {
              "$ref": "#/components/schemas/LeagueOut"
            },
            "title": "Leagues",
            "type": "array"
          },
          "teams": {
            "items": {
              "$ref": "#/components/schemas/TeamOut"
            },
            "title": "Teams",
            "type": "array"
          }
        },
        "required": [
          "leagues",
          "teams"
        ],
        "title": "TeamsOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "SportsDash",
    "version": "1.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/setup/follow": {
      "post": {
        "operationId": "setup_follow_api_setup_follow_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FollowRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Setup Follow"
      }
    },
...
    "/api/setup/teams/{league_id}": {
      "get": {
        "operationId": "setup_teams_api_setup_teams__league_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "league_id",
            "required": true,
            "schema": {
              "title": "League Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogTeamsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Setup Teams"
      }
    },
...
  }
}